♻️ CourseSearchedPlaceItem 분리
