+++ frontend/src/components/Course/CourseSearchedPlaceItem/CourseSearchedPlaceItem.typs.ts
@@ -1,0 +1,14 @@
+export interface SearchedPlace {
+  address_name: string;
+  category_group_code: string;
+  category_group_name: string;
+  category_name: string;
+  distance: string;
+  id: string;
+  phone: string;
+  place_name: string;
+  place_url: string;
+  road_address_name: string;
+  x: string;
+  y: string;
+}
