+++ frontend/src/hooks/TreeMap/useTreeMap.ts
@@ -1,21 +1,7 @@
 import { useCallback, useEffect, useRef, useState } from 'react';
+import { SearchedPlace } from '@/components/Course/CourseSearchedPlaceItem/CourseSearchedPlaceItem.typs';
 import { DEFAULT_LATITUDE, DEFAULT_LONGITUDE } from '@/constants/map';
 import treeImage from '@/assets/tree.png';
-
-interface SearchedPlace {
-  address_name: string;
-  category_group_code: string;
-  category_group_name: string;
-  category_name: string;
-  distance: string;
-  id: string;
-  phone: string;
-  place_name: string;
-  place_url: string;
-  road_address_name: string;
-  x: string;
-  y: string;
-}
 
 const { kakao } = window;
