+++ frontend/src/pages/Course/Course.tsx
@@ -1,8 +1,8 @@
 import { useState } from 'react';
 import { useEffect } from 'react';
-import { Link } from 'react-router-dom';
 import { IoIosSearch } from '@react-icons/all-files/io/IoIosSearch';
 import InputComboBox from '@/components/_common/InputComboBox/InputComboBox';
+import CourseSearchedPlaceItem from '@/components/Course/CourseSearchedPlaceItem/CourseSearchedPlaceItem';
 import useTreeMap from '@/hooks/TreeMap/useTreeMap';
 import { formatAddress } from '@/utils/formatAddress';
 import * as S from './Course.css';
@@ -54,21 +54,7 @@ const Course = () => {
             <h1 className={S.Title}>{currentCity} 주변 가볼 만한 곳</h1>
             <p className={S.SubTitle}>장소를 누르면 카카오맵으로 연결됩니다.</p>
           </div>
-          {searchedPlaceList &&
-            searchedPlaceList.map((place) => (
-              <Link to="http://place.map.kakao.com/784414359" target="_blank">
-                <div key={place.id} className={S.PlaceItemBox}>
-                  <img src={''} className={S.PlaceImage} alt="장소 썸네일" />
-                  <div className={S.PlaceContentBox}>
-                    <p className={S.PlaceName}>{place.place_name}</p>
-                    <div className={S.PlaceContentDetailBox}>
-                      <p className={S.PlaceAddress}>{place.road_address_name}</p>
-                      <p className={S.PlacePhone}>{place.phone}</p>
-                    </div>
-                  </div>
-                </div>
-              </Link>
-            ))}
+          {searchedPlaceList && searchedPlaceList.map((place) => <CourseSearchedPlaceItem place={place} />)}
         </section>
       </div>
     </div>
